@@ -50,8 +50,8 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": ">=14.15.0",
-    "npm": ">=6.14.8"
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
   },
   "scripts": {
     "commit": "git-cz",
