@@ -14,7 +14,7 @@ import {
   AttrSyntax,
   bindingCommand,
   IPlatform,
-  IAttrSyntaxTransformer,
+  IAttrMapper,
 } from '@aurelia/runtime-html';
 
 import type {
@@ -48,13 +48,13 @@ export class TranslationParametersBindingInstruction {
 export class TranslationParametersBindingCommand implements BindingCommandInstance {
   public readonly bindingType: BindingType.BindCommand = BindingType.BindCommand;
 
-  public static get inject() { return [IAttrSyntaxTransformer]; }
-  public constructor(private readonly t: IAttrSyntaxTransformer) {}
+  public static get inject() { return [IAttrMapper]; }
+  public constructor(private readonly m: IAttrMapper) {}
 
   public build(info: ICommandBuildInfo): TranslationParametersBindingInstruction {
     let target: string;
     if (info.bindable == null) {
-      target = this.t.map(info.node, info.attr.target)
+      target = this.m.map(info.node, info.attr.target)
         // if the transformer doesn't know how to map it
         // use the default behavior, which is camel-casing
         ?? camelCase(info.attr.target);
